Second round of protocol changes
A shot at making the actual protocol deal in trees, and not just
files.

The concept of pack SHA1s is gone completely.  We always talk about
packs in terms of a starting set of references and a finishing set.
When we download things in blocks, it is based on a conceptual "pack"
which is just the objects in the pack, back to back, with no overhead.

We sacrifice the ability to transfer partial objects in return for
being able to send real (shallow) git packs on the wire for responses;
those can basically be unpacked and stored, once all previous packs
are in place.

The "index" request has been modified; no longer sending a real index,
it lists the position of objects within the pack and their length.  It
is also moved to the "Data-oriented Messages" section, as it is
relatively heavy on network transfer.  However, if you are only
downloading from peers who are trading blocks of the same pack, then
you should not need to use this request.

The "request" request now supports sending a list of objects, and a
list of head refs that you claim to have, so that the result can be a
nice pack.

There is a new "blocks" request which should return you a quick bitmap
of which parts of the pack file the remote end has.

There may still be some gaps remaining; for instance, we still haven't
defined the sort order of objects within a pack.