"interval" negotiation - "valid" and "expires"
Instead of the tracker just blindly picking an update interval, allow
the peer to say how long they expect to be in the swarm, and the
server to then reply with an expires time.  The tracker can always
tell the client to back off with a 503 response (add to comment).